+++ src/oms_sensemaking/models/logs.py
@@ -1,10 +1,8 @@
 """Module for custom logging database objects"""
 
-import enum
-import logging
 import uuid
-from datetime import datetime
 
+from oms_sdk.generated.generated_graphql_client.enums import Action, ObjectType
 from sqlalchemy import UUID, Enum, String
 from sqlalchemy.dialects.postgresql import TEXT
 from sqlalchemy.orm import Mapped, mapped_column
@@ -13,22 +11,12 @@ from oms_sensemaking.models.base import (
     BaseORM,
     CreatedAuditMixin,
     SecurityMarkingMixin,
-    UtcDateTime,
-    utcnow_with_timezone,
 )
 
 
-class LogLevel(enum.Enum):
-    DEBUG = logging.DEBUG
-    INFO = logging.INFO
-    WARNING = logging.WARNING
-    ERROR = logging.ERROR
-    CRITICAL = logging.CRITICAL
+class AuditLogError(BaseORM, SecurityMarkingMixin, CreatedAuditMixin):
 
-
-class LogRecord(BaseORM, SecurityMarkingMixin, CreatedAuditMixin):
-
-    log_id: Mapped[int] = mapped_column(
+    id: Mapped[int] = mapped_column(
         UUID(as_uuid=True),
         nullable=False,
         primary_key=True,
@@ -36,16 +24,20 @@ class LogRecord(BaseORM, SecurityMarkingMixin, CreatedAuditMixin):
         comment="The unique ID of the Log.",
         default=uuid.uuid4
     )
-    timestamp: Mapped[datetime] = mapped_column(
-        UtcDateTime,
-        unique=False,
+    object_id: Mapped[uuid.UUID] = mapped_column(
+        UUID(as_uuid=True),
         nullable=False,
-        insert_default=utcnow_with_timezone,
-        comment="The time the log occurred.",
-    )
-    level: Mapped[LogLevel] = mapped_column(Enum(LogLevel), nullable=False, comment="Log Record Level")
+        comment="The unique id of the object in OMS.")
+    object_type: Mapped[str] = mapped_column(
+        Enum(ObjectType),
+        nullable=False,
+        comment="The type of object that the event was triggered on.")
+    event_type: Mapped[str] = mapped_column(
+        Enum(Action),
+        nullable=False,
+        comment="They type of event (e.g. create, update, or delete).")
     module_name: Mapped[str] = mapped_column(String, nullable=False, comment="Name of the producing thread")
     message: Mapped[str] = mapped_column(TEXT, nullable=False, comment="Log message")
     exc_text: Mapped[str] = mapped_column(TEXT, nullable=True, comment="Exception text")
 
-    __tablename__: str = "log_record"
+    __tablename__: str = "audit_log_error"