+++ src/oms_sensemaking/nlp/corenlp_client.py
@@ -21,7 +21,7 @@ class CoreNlpClient:
             self._props = {"annotators": "tokenize, pos, lemma, ner, depparse, relation", "outputFormat": "text"}
         else:
             self._props = props
-        LOGGER.debug(f"CoreNLP Client 'properties' is set to {self._props}")
+        LOGGER.warning(f"CoreNLP Client 'properties' is set to {self._props}")
 
         self.hostname = hostname
         self.url = f"http://{self.hostname}/?properties={self._props}"