+++ tests/geospatial/test_controller.py
@@ -183,7 +183,7 @@ def test_geo_controller_config(
 
     # Ensure that the default aircraft config is used
     instance.submit.assert_called_with(mock_geo_controller._registry["geo"].execute, track, default_aircraft_config)
-    mock_geo_controller.oms_crud_tool.get_source.assert_called_with(source_id=source_id)
+    mock_geo_controller.oms_crud_tool.get_source.assert_called_with(source_id=str(source_id))
 
     # Test Aircraft with relevant provider
     mock_geo_controller.oms_crud_tool.get_source = mock.MagicMock(return_value=source_relevant_provider)
@@ -207,4 +207,4 @@ def test_geo_controller_config(
     instance.submit.assert_called_with(mock_geo_controller._registry["geo"].execute, track, provider_1_aircraft_config)
 
     # Ensure provider id is fetched from most recent observation
-    mock_geo_controller.oms_crud_tool.get_source.assert_called_with(source_id=source_id)
+    mock_geo_controller.oms_crud_tool.get_source.assert_called_with(source_id=str(source_id))