+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Expanded further on Mil Symbol configuration to encompass various symbol Id formats found high side
+- Added AuditLogError database table to store errors
 
 ### Changed